Created battle and enemy movement

+++ engine.py
@@ -7,6 +7,7 @@ import time
 import numpy
 import game_screen
 import game_info
+import random
 from map_creator import generate_map
 from game_screen import resolve_screen
 from discord.ext import commands
@@ -28,19 +29,21 @@ monster_gallery = [
                 "s",
                 "p"]
 battle_commands = ["slash",
-                "cast"]
+                "cast",
+                "kill"]
 active_command_list = ["logout",
                 "login",
                 "end mission",
                 "exit",
                 "pause",
                 "print",
-                "enter"]
+                "enter",
+                "help"]
 starting_command_list = [  "login",
                     "continue",
                     "new game",
-                    "test"
-                    ]
+                    "test",
+                    "help"]
 movement_list = ["n",
                 "e",
                 "s",
@@ -159,23 +162,30 @@ async def engine(text_input, user_id_info):
                         #await text_input.channel.purge(limit=defaultval)
                         for x in range(iteration):
                             #I want to move the player using:
-                            text = await move_player(mov_dir, str(text_input.author.id), text_input)
-                            enemy = await move_enemies(text_input)
-                            if enemy == "ebattle":
-                                await start_battle(enemy, text_input)
-                            elif text == "pbattle":
-                                await start_battle(text, text_input)
-                            else:
-                                await resolve_screen(text_input)
-                                if text != "null":
-                                    await text_input.channel.send(text)
+                            if out_of_battle(text_input):
+                                text = await move_player(mov_dir, str(text_input.author.id), text_input)
+                                enemy = await move_enemies(text_input)
+                                if enemy == "ebattle":
+                                    await start_battle(enemy, text_input)
+                                elif text == "pbattle":
+                                    await start_battle(text, text_input)
+                                else:
+                                    await resolve_screen(text_input)
+                                    if text != "null":
+                                        await text_input.channel.send(text)
                     else:
                         await text_input.channel.send(command)
                 else:
                     await text_input.channel.send("That is not a recognized command, " + username + ". Would you like some *help*, to *pause*, or even *end mission*?")
             else:
-                await text_input.channel.send("in battle!")
-                await battle_round(text_input)
+                is_battle_command = False
+                for battle_command in battle_commands:
+                    if str(text_input.content).lower() == battle_command:
+                        is_battle_command = True
+                if is_battle_command:
+                    await battle_round(text_input)
+                else:
+                    await text_input.channel.send("That is not a recognized command, " + username + ". Would you like some *help*, to *pause*, or even *end mission*?")
         else:
             if str(text_input.content).lower() == "next":
                 await game_screen.print_credits(text_input, False)
@@ -396,10 +406,110 @@ async def input_parse(raw_input):
 #|   |___ | | |   ||   |___ | ||_|| |  |   |    |   |    |   _   |  |   |  |   _   ||   |    |   | | | |   ||       ||   | | | |   ||   |_| |
 #|_______||_|  |__||_______||_|   |_|  |___|    |___|    |__| |__|  |___|  |__| |__||___|    |___| |_|  |__||______| |___| |_|  |__||_______|
 #-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+enemy_sight = {
+    "m" : 2,
+    "b" : 3,
+    "s" : 2,
+    "p" : 3,
+    "w" : 9,
+    "S" : 9,
+    "I" : 9,
+    "W" : 9,
+    "!" : 9  
+}
+entity_list = ["m","b","s","p","w","S","I","W","!","<","?","#"]
 
 async def move_enemies(raw_input):
-    await raw_input.channel.send("move enemies!")
-    return "NULL"
+    #First things first we want to load the map into an array
+    with open("active_" + str(raw_input.author.id) + ".txt") as f:
+        input_grid = f.readlines()
+        input_grid = [row.rstrip('\n') for row in input_grid]
+        room_width = len(input_grid)
+        room_height = int((len(input_grid[0]) + 1) / 2)
+        #Now that we have the grid let's remove the whitespaces
+        void = "-"
+        grid_array = [[void for i in range(room_height)] for j in range(room_width)]
+        for y in range(0,len(input_grid)-1):
+            for x in range(0,len(input_grid[0])-1):
+                if x % 2 == 0:
+                    grid_array[y][int(x/2)] = input_grid[y][x]
+    fname = "info_" + str(raw_input.author.id) + ".txt"
+    info_array = numpy.genfromtxt(fname, dtype=str, delimiter=",")
+    amount_of_enemies = 0
+    player_x = info_array[0][9]
+    player_y = info_array[0][10]
+    #We know that the player character is on entity #1, we need to know how many enemies there are.
+    #Luckily we know that they're sandwiched between the player and the door, so figure out where the door is, subtract 1 and that's the amount of enemies
+    for row in range(len(info_array)):
+        if info_array[row][1] == "<":
+            amount_of_enemies = int(info_array[row][0]) - 2
+    #Now that we have the amount of enemies, let's start cycling through them and updating everything.
+    for enemy in range(amount_of_enemies):
+        #First lets grab the units x and y
+        enemy_type = info_array[enemy + 2][1]
+        #check if the enemy is dead first
+        if enemy_type == "X":
+            continue
+        enemy_x = info_array[enemy + 2][9] #we add 2 to get back to the entity number
+        enemy_y = info_array[enemy + 2][10]
+        encounter_direction = 0
+        #Now lets see if the enemy is near the player
+        diff_x = abs(enemy_x - player_x)
+        diff_y = abs(enemy_y - player_y)
+        if diff_x <= enemy_sight[enemy_type] and diff_y <= enemy_type[enemy_type]:
+            #If within their sight, we want to grab which value is closest (if there is one)
+            if diff_x == diff_y:
+                x_or_y = random.randint(0,1)*2-1
+                y = 1
+                if x_or_y == y:
+                    if enemy_y < player_y:
+                        encounter_direction = 2
+                    else:
+                        encounter_direction = 0
+                else:
+                    if enemy_x < player_x:
+                        encounter_direction = 1
+                    else:
+                        encounter_direction = 3
+            elif diff_y < diff_x:
+                if enemy_y < player_y:
+                    encounter_direction = 2
+                else:
+                    encounter_direction = 0
+            else:
+                if enemy_x < player_x:
+                    encounter_direction = 1
+                else:
+                    encounter_direction = 3
+        #If we're not within sight, then we want to grab their path that they should be on
+        #At the end we'll encounter the space and then move from there
+        else:
+            #Screw the pathing for a bit and lets just move the enemies around randomly
+            encounter_direction = random.randint(0,3)
+            
+        encounter_char, new_x, new_y = encounter_space(encounter_direction,enemy_x, enemy_y, raw_input)
+        #Need to find out if the encountered area is an entity or not
+        for entity in range(len(entity_list)):
+            if encounter_char == entity_list[entity]:
+                return "null"
+        if encounter_char == "&":
+            #If we've hit the player, we're gonna have to enter battle!
+            #First update the info, then return the enemy battle tag
+            in_battle = 1
+            update_info = ["1", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", in_battle]
+            await game_info.force_update(raw_input, update_info)
+            update_info = [str(enemy + 2), "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", new_x, new_y, "NULL", "NULL", in_battle]
+            await game_info.force_update(raw_input, update_info)
+            return "ebattle"
+        else:
+            #otherwise we hit a space and lets update the map and info with that
+            update_info = [str(enemy + 2), "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", new_x, new_y, "NULL", "NULL", "NULL"]
+            await game_info.force_update(raw_input, update_info)
+            grid_array[enemy_y][enemy_x] = " "
+            grid_array[new_y][new_x] = enemy_type
+            file_name = "active_" + str(raw_input.author.id) + ".txt"
+            numpy.savetxt(file_name, grid_array, fmt='%s')
+    return "null"
 
 
 
@@ -454,7 +564,10 @@ async def encounter_space(direction, entity_x, entity_y, raw_input):
 
 async def start_battle(initiate, raw_input):
     #I'll have to use the raw input to get the x and y and of the battle through the info file
-    await raw_input.channel.send("start battle")
+    if initiate == "pbattle":
+        await raw_input.channel.send("player initiated battle")
+    elif initiate == "ebattle":
+        await raw_input.channel.send("enemy initiated battle")
     fname = "battle_" + str(raw_input.author.id) + ".txt"
     x = [1]
     numpy.savetxt(fname, x)
@@ -468,11 +581,89 @@ async def start_battle(initiate, raw_input):
 #| |_|   ||   _   |  |   |    |   |  |       ||   |___   |   |  | ||       ||       || | |   ||       |
 #|_______||__| |__|  |___|    |___|  |_______||_______|  |___|  |_||_______||_______||_|  |__||______| 
 #-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
+enemy_name_list = {
+    "m" : "nasty mouse",
+    "b" : "giant bat",
+    "s" : "slithering snake",
+    "p" : "terrible spider",
+    "w" : "9",
+    "S" : "9",
+    "I" : "9",
+    "W" : "9",
+    "!" : "9"
+}
 async def battle_round(raw_input):
     #This is where we'll reopen the battle file and continue
+    for battle_command in battle_commands:
+        if str(raw_input.content).lower() == battle_command:
+            #lets first find the enemy the player has encountered
+            fname = "info_" + str(raw_input.author.id) + ".txt"
+            entity_val = 0
+            entity_char = str(info_array[entity_val-1][1])
+            info_array = numpy.genfromtxt(fname, dtype=str, delimiter=",")
+            for row in range(len(info_array)):
+                if row > 0 and int(info_array[row][13]) == 1:
+                    entity_val = info_array[row][0]
+            #Let's hurt the player and such
+            total_player_attack = int(info_array[0][8]) + int(info_array[0][2])
+            total_player_health = int(info_array[0][7]) + int(info_array[0][3])
+            total_enemy_attack = int(info_array[entity_val-1][8]) + int(info_array[entity_val-1][2])
+            total_enemy_health = int(info_array[entity_val-1][7]) + int(info_array[entity_val-1][3])
+            output_string = "You swung your sword for " + str(total_player_attack) + " points of damage!"
+            await raw_input.channel.send(output_string)
+            if total_player_attack >= total_enemy_health:
+                output_string = "You killed the " + enemy_name_list[entity_char] + "!"
+                await raw_input.channel.send(output_string)
+                #then update the info file
+                in_battle = 0
+                update_info = ["1", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", "NULL", in_battle]
+                await game_info.force_update(raw_input, update_info)
+                update_info = [str(entity_val + 1), "X", "NULL", 0, "NULL", "NULL", "NULL", 0, "NULL", "NULL", "NULL", "NULL", "NULL", in_battle]
+                await game_info.force_update(raw_input, update_info)
+                fname = "battle_" + str(raw_input.author.id) + ".txt"
+                os.remove(fname)
+            else:
+                total_enemy_health -= total_player_attack
+                output_string = "The " + enemy_name_list[entity_char] + " attacked you for "+ str(total_enemy_attack) + " points of damage!"
+                await raw_input.channel.send(output_string)
+                total_player_health -= total_enemy_attack
+                if total_player_health <= 0:
+                    fname = "battle_" + str(raw_input.author.id) + ".txt"
+                    os.remove(fname)
+                    fname = "info_" + str(raw_input.author.id) + ".txt"
+                    os.remove(fname)
+                    fname = "active_" + str(raw_input.author.id) + ".txt"
+                    os.remove(fname)
+                    await raw_input.channel.purge(limit=defaultval)
+                    await raw_input.channel.send("You have died in the catacombs; Better luck next time!")
+                    await raw_input.channel.send("Would you like to start a *new game*?")
+                else:
+                    if total_player_health <= int(info_array[0][3]):
+                        new_player_health = total_player_health
+                        new_player_armor = 0
+                    elif total_player_health > int(info_array[0][3]):
+                        new_player_health = int(info_array[0][3])
+                        new_player_armor = total_player_health - int(info_array[0][3])
+                    #Now the same for enemies
+                    if total_enemy_health <= int(info_array[entity_val-1][3]):
+                        new_enemy_health = total_enemy_health
+                        new_enemy_armor = 0
+                    elif total_enemy_health > int(info_array[entity_val-1][3]):
+                        new_enemy_health = int(info_array[entity_val-1][3])
+                        new_enemy_armor = total_enemy_health - int(info_array[entity_val-1][3])
+                    #Now to save the health
+                    update_info = ["1", "NULL", "NULL", new_player_health, "NULL", "NULL", "NULL", new_player_armor, "NULL", "NULL", "NULL", "NULL", "NULL", "NULL"]
+                    await game_info.force_update(raw_input, update_info)
+                    update_info = [str(entity_val + 1), "NULL", "NULL", new_enemy_health, "NULL", "NULL", "NULL", new_enemy_armor, "NULL", "NULL", "NULL", "NULL", "NULL", "NULL"]
+                    await game_info.force_update(raw_input, update_info)
+
+async def out_of_battle(raw_input):
+    #This is where we'll reopen the battle file and continue
+    out_of_battle = True
     fname = "battle_" + str(raw_input.author.id) + ".txt"
-    os.remove(fname)
+    if os.path.isfile(fname):
+                out_of_battle = False
+    return out_of_battle
 
 #-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 #                 __    _  _______  __   __  _______
